+++ src/Config.js
@@ -346,6 +346,13 @@ const config = {
 
 					// https://www.npmjs.com/package/partialify
 					options : {}
+				},
+
+				{
+					name: 'vueify',
+
+				    // https://github.com/vuejs/vueify#usage
+				    options: {}
 				}
 			],