+++ test-app/resources/assets/js/main.js
@@ -1,0 +1,10 @@
+// main.js
+var Vue = require( 'vue' )
+var App = require( './components/test.vue' )
+
+new Vue({
+	el : 'body',
+	components : {
+		app : App
+	}
+})