+++ changelog.md
@@ -9,6 +9,8 @@
 - If not tasks defined in the user's gulpfile, just a message is shown instead of an exception
 - Removed coffeescript support, no longer needed
 - Added new config map for `appPaths` to support all ColdBox conventions
+- Vueify core support view browserify for Vue.js components
+- Included a new `template` folder which helps setting up new elixir based projects
 
 ## 1.0.0 
 - Initial Port to ColdBox